api/requirements.txt
fastapi
python-multipart
pydantic
pydantic-settings
supabase
python-dotenv
resend
groq
